Moved to header utils
