+++ src/main/java/org/folio/event/handler/InventoryCreateAsyncRecordHandler.java
@@ -1,8 +1,8 @@
 package org.folio.event.handler;
 
-import static org.folio.event.util.KafkaEventUtil.TENANT_NOT_SPECIFIED_MSG;
-import static org.folio.event.util.KafkaEventUtil.extractTenantFromHeaders;
-import static org.folio.kafka.KafkaHeaderUtils.kafkaHeadersToMap;
+import static org.folio.util.HeaderUtils.TENANT_NOT_SPECIFIED_MSG;
+import static org.folio.util.HeaderUtils.extractTenantFromHeaders;
+import static org.folio.util.HeaderUtils.getHeaderMap;
 import static org.folio.util.HeaderUtils.prepareHeaderForTenant;
 
 import java.util.Map;
@@ -67,7 +67,7 @@ public abstract class InventoryCreateAsyncRecordHandler extends BaseAsyncRecordH
         return Future.succeededFuture();
       }
 
-      var headers = new CaseInsensitiveMap<>(kafkaHeadersToMap(kafkaConsumerRecord.headers()));
+      var headers = getHeaderMap(kafkaConsumerRecord.headers());
       return getCentralTenantId(headers)
         .compose(centralTenantId -> processInventoryCreationEventIfNeeded(resourceEvent, centralTenantId, headers))
         .onSuccess(v -> log.info("handle:: '{}' event for '{}' processed successfully", eventType, inventoryEventType.getTopicName()))
