+++ src/test/java/org/folio/event/KafkaEventUtilTest.java
@@ -1,12 +1,12 @@
 package org.folio.event;
 
+import java.util.List;
+
 import io.vertx.kafka.client.producer.KafkaHeader;
 import io.vertx.kafka.client.producer.impl.KafkaHeaderImpl;
-import org.folio.event.util.KafkaEventUtil;
+import org.folio.util.HeaderUtils;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-
-import java.util.List;
 
 public class KafkaEventUtilTest {
 
@@ -14,7 +14,7 @@ public class KafkaEventUtilTest {
   void successRetrieveHeader() {
     String expValue = "diku";
     KafkaHeader kafkaHeader = new KafkaHeaderImpl("x-okapi-header", expValue);
-    String actValue = KafkaEventUtil.extractValueFromHeaders(List.of(kafkaHeader), "x-okapi-header");
+    String actValue = HeaderUtils.extractValueFromHeaders(List.of(kafkaHeader), "x-okapi-header");
     Assertions.assertEquals(expValue, actValue);
   }
 
@@ -22,7 +22,7 @@ public class KafkaEventUtilTest {
   void ifNoFoundThenReturnNull() {
     String expValue = "diku";
     KafkaHeader kafkaHeader = new KafkaHeaderImpl("x-okapi-header1", expValue);
-    String actValue = KafkaEventUtil.extractValueFromHeaders(List.of(kafkaHeader), "x-okapi-header");
+    String actValue = HeaderUtils.extractValueFromHeaders(List.of(kafkaHeader), "x-okapi-header");
     Assertions.assertNull(actValue);
   }
 }
